Add minimal repro for ionic navigation bug
To reproduce, install dependencies and run the app (`ionic serve`).

Type something in the input, submit the form, then click on the link. You'll notice that the input has the same value as the previous form view. Navigate back twice. Now you'll notice that the input has lost its value.

It seems that ionic router uses the wrong instance of the view, re-using the old instance for the new, unique URL and clobbering the one in the history stack.